// types/Track.ts
export type Track = {
  name?: string;
  album_id?: string;
  artist_name?: string;
  artist_id?: string;
  comments?: JSON;
  created_at?: string;
  rating?: number;
  duration_ms?: number;
  genre?: string;
  spotify_track_id?: string;
  spotify_uri?: string;
  track_id?: string;
  track_name: string;
};